requirements.txt
selenium
pytest
pytest-html
pytest-xdist
allure-pytest